--- a/ecmascript/transforms/src/compat/es2020/opt_chaining.rs
+++ b/ecmascript/transforms/src/compat/es2020/opt_chaining.rs
@@ -6,6 +6,7 @@ use std::{fmt::Debug, iter::once, mem};
 use swc_common::{Spanned, DUMMY_SP};
 use swc_ecma_ast::*;
 use swc_ecma_transforms_macros::fast_path;
+use swc_ecma_utils::alias_if_required;
 use swc_ecma_visit::{noop_fold_type, Fold, FoldWith, Node, Visit};
 
 pub fn optional_chaining() -> impl Fold {
@@ -330,9 +331,8 @@ impl OptChaining {
                 type_args,
                 ..
             }) => {
-                let obj = *obj;
                 let obj_span = obj.span();
-                let is_super_access = match obj {
+                let is_super_access = match *obj {
                     Expr::Member(MemberExpr {
                         obj: ExprOrSuper::Super(..),
                         ..
@@ -340,9 +340,35 @@ impl OptChaining {
                     _ => false,
                 };
 
-                let (left, right, alt) = match obj {
-                    Expr::Ident(..) => (Box::new(obj.clone()), Box::new(obj), e.expr),
+                let (left, right, alt) = match *obj {
+                    Expr::Ident(..) => (obj.clone(), obj, e.expr),
                     _ => {
+                        let this_as_super;
+                        let (this_obj, aliased) = alias_if_required(
+                            match &*obj {
+                                Expr::Member(m) => match &m.obj {
+                                    ExprOrSuper::Super(s) => {
+                                        this_as_super = Expr::This(ThisExpr { span: s.span });
+                                        &this_as_super
+                                    }
+                                    ExprOrSuper::Expr(obj) => &**obj,
+                                },
+                                _ => &*obj,
+                            },
+                            "_obj",
+                        );
+                        let obj = if !is_super_access && aliased {
+                            self.vars.push(VarDeclarator {
+                                span: obj_span,
+                                definite: false,
+                                name: Pat::Ident(this_obj.clone()),
+                                init: Some(obj),
+                            });
+
+                            Box::new(Expr::Ident(this_obj.clone()))
+                        } else {
+                            obj
+                        };
                         let i = private_ident!(obj_span, "ref");
                         self.vars.push(VarDeclarator {
                             span: obj_span,
@@ -356,7 +382,7 @@ impl OptChaining {
                                 span: DUMMY_SP,
                                 left: PatOrExpr::Pat(Box::new(Pat::Ident(i.clone()))),
                                 op: op!("="),
-                                right: Box::new(obj),
+                                right: obj,
                             })),
                             Box::new(Expr::Ident(i.clone())),
                             Box::new(Expr::Call(CallExpr {
@@ -367,11 +393,10 @@ impl OptChaining {
                                     prop: Box::new(Expr::Ident(Ident::new("call".into(), span))),
                                     computed: false,
                                 }))),
-                                // TODO;
                                 args: once(if is_super_access {
                                     ThisExpr { span }.as_arg()
                                 } else {
-                                    i.as_arg()
+                                    this_obj.as_arg()
                                 })
                                 .chain(args)
                                 .collect(),