swc_ecma_transforms:
 - Fix `this` in optional chaining expression. (https://github.com/Brooooooklyn/swc-node/issues/62)
 - Optimize typescript stripper
 - Optimize fixer


---

Note: Tasks to do are determined by looking at the perf report.

---

### Benchmark


#### Baseline
```
test full_es2016                  ... bench:   2,305,810 ns/iter (+/- 286,352)
```

#### After removing map_with_mut in typescript::strip#visit_mut_expr

```
test full_es2016                  ... bench:   1,975,390 ns/iter (+/- 168,977)
```


#### After reducing the number of branches in fixer

I added some transform-only benchmarks.

```
test full_es2016                  ... bench:   1,825,710 ns/iter (+/- 27,324)
test transforms_es2016            ... bench:     877,426 ns/iter (+/- 80,775)
test parser                       ... bench:     776,628 ns/iter (+/- 19,839)
test codegen_es2016               ... bench:     137,076 ns/iter (+/- 1,126)
```